Refactor to avoir code duplication
